translation/requirements.txt
openai==1.82.0
python-dotenv==1.0.1
langchain_community==0.3.24
nltk==3.9.1
transformers==4.52.3